Finish renaming and cleaning __init__.py
