+++ openpathsampling/experimental/simstore/dict_serialization_helpers.py
