+++ openpathsampling/experimental/storage/__init__.py
@@ -1,8 +1,11 @@
 # TODO: improve the imports here
 from . import snapshots
 from . import monkey_patches
-from . import dict_serialization_helpers
 
 from .ops_storage import (
     Storage, OPSClassInfoContainer, ops_schema, ops_class_info
+)
+
+from .monkey_patches import (
+    monkey_patch_saving, monkey_patch_loading, monkey_patch_all
 )
